% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b31crvl.R
\name{b31crvl}
\alias{b31crvl}
\title{ASME B31G. Basic computer program CRVL.BAS}
\usage{
b31crvl(maop, d, wth, smys, def = 0.72, depth, l)
}
\arguments{
\item{maop}{maximum allowable operating pressure - \emph{MAOP}, [\emph{PSI}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{d}{nominal outside diameter of pipe, [\emph{inch}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{wth}{nominal wall thickness of pipe, [\emph{inch}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{smys}{specified minimum yield of stress (\emph{SMYS}) as a characteristics of
steel strength, [\emph{PSI}]. Type: \code{\link[checkmate]{assert_double}}.}

\item{def}{appropriate (combined) design factor from
\href{https://law.resource.org/pub/us/cfr/ibr/002/asme.b31.4.2002.pdf}{
ASME B31.4},
\href{https://law.resource.org/pub/us/cfr/ibr/002/asme.b31.8.2003.pdf}{
ASME B31.8}, or
\href{https://www.asme.org/codes-standards/find-codes-standards/b31-11-slurry-transportation-piping-systems}{
 ASME B31.11}, [].
Type: \code{\link[checkmate]{assert_double}}.}

\item{depth}{measured maximum depth of the corroded area, [\emph{inch}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{l}{measured maximum longitudinal length of the corroded area, [\emph{inch}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
Object of \emph{S3}-class \emph{crvl} which is a \emph{data.frame} with
 the next numeric columns:

 \describe{
   \item{maop}{
     maximum allowable operating pressure - \emph{MAOP}, [\emph{PSI}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{d}{
     nominal outside diameter of pipe, [\emph{inch}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{wth}{
     nominal wall thickness of pipe, [\emph{inch}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{smys}{
     specified minimum yield of stress (\emph{SMYS}) as a characteristics of
     steel strength, [\emph{PSI}].
     Type: \code{\link[checkmate]{assert_double}}.
    }
   \item{def}{
     appropriate (combined) design factor from
     \href{https://law.resource.org/pub/us/cfr/ibr/002/asme.b31.4.2002.pdf}{
     ASME B31.4},
     \href{https://law.resource.org/pub/us/cfr/ibr/002/asme.b31.8.2003.pdf}{
     ASME B31.8},
     or
     \href{https://www.asme.org/codes-standards/find-codes-standards/b31-11-slurry-transportation-piping-systems}{ASME B31.11},
     []. Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{depth}{
     measured maximum depth of the corroded area, [\emph{inch}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{l}{
     measured maximum longitudial length of corroded area, [\emph{inch}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{status}{
   operational status of pipe:
     \emph{1} - excellent,
     \emph{2} - monitoring is recommended,
     \emph{3} - alert! replace the pipe immediately!
     Type: \code{\link[checkmate]{assert_numeric}}.
   }
   \item{design_pressure}{
     design pressure of pipe, [\emph{PSI}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{safe_pressure}{
     safe maximum pressure for the corroded area, [\emph{PSI}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{pressure_exceeding}{
     whether operator's action is required to reduce \emph{MOAP} lower than
     the maximum safe pressure of the corroded area.
     Type: \code{\link[checkmate]{assert_logical}}.
   }
   \item{allowed_corrosion_depth}{
     allowable depth of the corroded area, [\emph{inch}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{A}{
     intermediate factor related to the geometry of the corroded area, [].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{allowed_corrosion_length}{
     allowable length of the corroded area, [\emph{inch}].
     Type: \code{\link[checkmate]{assert_double}}.
   }
   \item{AP}{
     another intermediate factor related to the geometry of the corroded
     area, []. Type: \code{\link[checkmate]{assert_double}}.
   }
 }
}
\description{
Imitation of \emph{CVRL.BAS} computer program presented in
 \href{https://law.resource.org/pub/us/cfr/ibr/002/asme.b31g.1991.pdf}{
 ASME B31G-1991} \emph{Appendix A} for determining allowable length and
 allowable operating pressure
}
\details{
Columns \emph{maop}, \emph{d}, \emph{wth}, \emph{smys}, \emph{def},
  \emph{depth}, \emph{l} in the output \emph{data.frame} come from
  function's input, other columns are calculated.

  For univariate case (when lengths of all input vectors are one) messages
  that imitate \emph{CRVL.BAS} console output are printed.
}
\examples{
library(pipenostics)

## Further examples are inspired by those used in Appendix A of
## ASME B31G-1991 to verify correct entry of CRVL.BAS source code:

## Standard example 1
b31crvl(
  maop = 910, d = 30, wth = .438, smys = 52000, def  = .72, depth = .1,
  l = 7.5
)

## Standard example 2
b31crvl(
  maop = 400, d = 20, wth = .25, smys = 35000, def  = 0.5, depth = 0.18,
  l = 10
)

## Standard example 3
b31crvl(
  maop = 910, d = 24, wth = .432, smys = 52000, def  = .72, depth = 0.13,
  l = 30
)

## Standard example 4
b31crvl(
  maop = 910, d = 24, wth = .432, smys = 52000, def  = .72, depth = .3,
  l = 30
)

## Standard example 5
b31crvl(
  maop = 731, d = 24, wth = .281, smys = 52000, def  = 0.72, depth = 0.08,
  l = 15
)

## Standard example 6
b31crvl(
  maop = 1e3, d = 36, wth = .5, smys = 52000, def  = 0.72, depth = 0.41,
  l = 100
)

## Standard example 8
b31crvl(
  maop = 877, d = 12.625, wth = .5, smys = 35000, def  = .4, depth = .035,
  l = 3
)

## Standard example 9
b31crvl(
  maop = 790, d = 24, wth = .5, smys = 42000, def  = .5, depth = .125, l = 12
)

## TEST #1
b31crvl(
  maop = 790, d = 24, wth = .5, smys = 42000, def  = .5, depth = .179, l = 12
)

## TEST #1A
b31crvl(
  maop = 790, d = 24, wth = .5, smys = 42000, def  = .5, depth = .179,
  l = 12.182
)

## TEST #1B
b31crvl(
  maop = 790, d = 24, wth = .5, smys = 42000, def  = .5, depth = .180,
  l = 12.182
)

## TEST #2
b31crvl(
  maop = 790, d = 24, wth = .5, smys = 42000, def  = .5, depth = .179,
  l = 12.297
)

## All examples at once:
data(b31gdata)
with(b31gdata, b31crvl(maop, d, wth, smys, def, depth, l))

}
\references{
\href{https://law.resource.org/pub/us/cfr/ibr/002/asme.b31g.1991.pdf}{
 ASME B31 G-1991}.
 Manual for determining the remaining strength of corroded pipelines. A
 supplement to \emph{ASME B31G} code for pressure piping.
}
\seealso{
Other ASME B31G: 
\code{\link{b31gacd}()},
\code{\link{b31gacl}()},
\code{\link{b31gafr}()},
\code{\link{b31gdep}()},
\code{\link{b31gops}()},
\code{\link{b31gsap}()}
}
\concept{ASME B31G}
