% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b36dwthv.R
\name{b36dwthv}
\alias{b36dwthv}
\title{ASME B36.10M. Verify the alignment between nominal pipe diameter and wall
 thickness}
\usage{
b36dwthv(d, wth)
}
\arguments{
\item{d}{nominal (outside) diameter of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{wth}{nominal wall thickness of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
Verification status. Type: \code{\link[checkmate]{assert_logical}}.
}
\description{
Check the correspondence inside nominal diameter/wall thickness value pair
 for pipes manufactured in accordance with regulatory standards.

 Verification is performed using reference values from the
 \code{\link{b36pipedata}}-table. So, the result is \code{TRUE} if \code{d}-
 and \code{w}-value combination can be found in rows of this dataset.
}
\examples{
 library(pipenostics)

 # All pipes in test best of district heating network have strictly
 # consistent diameters and wall thiknesses:

 all(b36dwthv(m325nxdata[["d"]], m325nxdata[["wth"]]))

}
\seealso{
Other utils: 
\code{\link{b36mass}()},
\code{\link{b36pipedata}},
\code{\link{geoarea}()},
\code{\link{meteos}()},
\code{\link{mgtdhid}()}
}
\concept{utils}
