% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b36mass.R
\name{b36mass}
\alias{b36mass}
\alias{b36d}
\alias{b36wth}
\title{ASME B36.10M. Estimate pipe mass}
\usage{
b36mass(d, wth, len = 1, rho = 7.85, origin = NULL)

b36d(wth, mass, len = 1, rho = 7.85)

b36wth(d, mass, len = 1, rho = 7.85)
}
\arguments{
\item{d}{nominal outside diameter of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}}

\item{wth}{nominal wall thickness of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{len}{pipe length, [\emph{m}]. Type: \code{\link[checkmate]{assert_double}}.}

\item{rho}{mass density of pipe material, [\emph{g/cm³}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{origin}{identifier for the information origin regarding the specifications of pipe.
Type: \code{\link[checkmate]{assert_integer}}.}

\item{mass}{actual mass of pipe, [\emph{kg}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
\describe{
  \item{for \code{\link{b36mass}}}{pipe mass, [\emph{kg}]}
  \item{for \code{\link{b36wth}}}{pipe wall thickness, [\emph{mm}]}
  \item{for \code{\link{b36d}}}{outside diameter of pipe, [\emph{mm}]}
 }
}
\description{
Estimate mass or geometric specifications of the manufactured pipes
}
\details{
The mass of a one-meter pipe segment is determined by linear interpolation
  based on the tabular data provided in the specified origins (see
  \code{\link{b36pipedata}}). If the \code{origin} of the initial data is
  \code{NULL} (default), the mass of the pipe is calculated using a formula:

  \deqn{M = 10^{-3} \pi \rho w \left(d - w \right ) \cdot l}

  where
  \itemize{
     \item \eqn{M} - mass of a pipe, [\emph{kg}]
     \item \eqn{\rho} - mass density of pipe material, [\emph{g/cm³}]
     \item \eqn{w} - nominal wall thickness of the manufactured pipe,
                     [\emph{mm}]
     \item \eqn{d} - nominal outside diameter of the manufactured pipe,
                     [\emph{mm}]
     \item \eqn{l} - actual pipe length, [\emph{m}]
  }
For \code{origin} in \code{c(1, 2, 4, 5)} the values provided in the
\code{\link{b36pipedata}} match the calculated values obtained from the
formula with an accuracy of 1 \%.

Inverse calculations \code{\link{b36wth}} and \code{\link{b36d}} are
performed using algebraically derived formulas.
}
\examples{
library(pipenostics)

# Since some specification origins provide the mass of a one-meter pipe
# segment taking into account possible deviations during its production
# process, when the user specifies the origin ID directly, the mass values
# are calculated using linear interpolation:
b36mass(68, 13, rho = 7.9, origin = 7L)

# The discrepancy with the calculations based on the formula can be more
# than 7 \%:
b36mass(68, 13, rho = 7.9, origin = NULL)

# For origins which are ASME B36 standards such differences should be
# minimal:
b36mass(965, 10.31, origin = 1L) - b36mass(965, 10.31, origin = NULL)

# The calculations of diameter and wall thickness are straightforward and
# use only inverse formulas without origin references:
b36d(10.31, 242.74)

b36wth(965, 242.74)

}
\seealso{
Other ASME B36.10M: 
\code{\link{b36dwthv}()},
\code{\link{b36pipedata}}
}
\concept{ASME B36.10M}
