% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowls.R
\name{flowls}
\alias{flowls}
\title{List all possible flow paths in district heating network}
\usage{
flowls(sender = "A", acceptor = "B", use_cluster = FALSE)
}
\arguments{
\item{sender}{identifier of the node which heat carrier flows out.
Type: any type that can be painlessly coerced to character by
\code{\link{as.character}}.}

\item{acceptor}{identifier of the node which heat carrier flows in. According to topology
of test bench considered this identifier should be unique.
Type: any type that can be painlessly coerced to character by
\code{\link{as.character}}.}

\item{use_cluster}{utilize functionality of parallel processing on multi-core CPU.
Type: \code{\link[checkmate]{assert_flag}}.}
}
\value{
Named \code{\link{list}} that contains integer vectors as its elements. The
 name of each element in the \code{\link{list}} is the name of
 \code{acceptor} associated with terminal node of district heating network.
 Each vector in the \code{\link{list}} represents an ordered sequence of
 indexes in \code{acceptor} that enumerates incoming edges from starting node
 to terminal one. The length of returned \code{link{list}} is equal to number
 of terminal nodes for topology considered.
 Type: \code{\link[checkmate]{assert_list}}.
}
\description{
Find and list all possible paths of heat carrier flow (water) in the given
 topology of district heating system.
}
\details{
Only branched topology without cycles is considered where no more than one
incoming edge exists for every \code{acceptor} node. For instance,
\code{\link{m325nxdata}} has permitted topology.

Though input arguments are natively vectorized their individual values
all relate to common part of district heating network, i.e. associated with
common object. It is due to isomorphism between vector representation and
directed graph of this network. For more details of isomorphic topology
description see \code{\link{m325nxdata}}.
}
\examples{
library(pipenostics)

# Find path from A to B in trivial line topology:
flowls("A", "B")

# More complex example with two terminal nodes D and E:
flowls(c("A", "B", "B"), c("B", "D", "E"))

# All possible flow paths in test bench illustrated in `?m325nxdata`:
all_paths <- list(
  c(12, 13, 11, 8, 4, 1),  # hereinafter indexes of acceptor nodes
  c(12, 13, 11, 8, 4, 2),
  c(12, 13, 11, 8, 6, 5,  3),
  c(12, 13, 11, 8, 6, 7),
  c(12, 13, 11, 8, 6, 9),
  c(12, 13, 11, 10),
  c(12, 13, 14, 15),
  c(12, 13, 16, 17),
  c(12, 13, 16, 18, 20, 19),
  c(12, 13, 16, 18, 20, 21),
  c(12, 13, 16, 18, 22, 24),
  c(12, 13, 16, 18, 22, 25),
  c(12, 13, 16, 18, 20, 23, 26)
)

# find those paths:
with(m325nxdata, flowls(sender, acceptor))

}
\seealso{
\code{\link{m325nxdata}} for example of topology of district heating
 system
}
