% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{loss_flux}
\alias{loss_flux}
\alias{flux_loss}
\title{Convert heat flux to specific heat loss power}
\usage{
loss_flux(x, d = 720)

flux_loss(x, d = 720)
}
\arguments{
\item{x}{value of
\itemize{
  \item \emph{heat flux}, [\emph{W/m²}], for \code{loss_flux(x, d, wth)}
  \item \emph{specific heat loss power}, [\emph{kcal/m/h}], for
  \code{flux_loss(x, d, wth)(x)}
}
Type: \code{\link[checkmate]{assert_double}}.}

\item{d}{nominal (outside) diameter of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
Value of
 \itemize{
   \item \emph{specific heat loss power}, [\emph{kcal/m/h}], for
         \code{loss_flux(x, d, wth)}
   \item \emph{heat flux}, [\emph{W/m²}], for \code{flux_loss(x, d, wth)(x)}
 }
 Type: \code{\link[checkmate]{assert_double}}.
}
\description{
Convert \href{https://en.wikipedia.org/wiki/Heat_flux}{heat flux}
 measured for a cylindrical steel pipe to \emph{specific heat loss power} of
 pipe.
}
\examples{
library(pipenostics)

# Consider pipes with nominal (outside) diameters:
d <- c(998, 1395)  # [mm]

# Then maximum possible normative neat loss according (Minenergo-325) is
loss_max <- c(218, 1040)  # [kcal/m/h]

# The appropriate flux is
flux <- flux_loss(loss_max, d)
print(flux)

stopifnot(all.equal(loss_flux(flux, d), loss_max, tolerance = 1e-5))

}
\seealso{
Other Measurement Unit Converter: 
\code{\link{c_k}()},
\code{\link{f_k}()},
\code{\link{inch_mm}()},
\code{\link{k_c}()},
\code{\link{kgf_mpa}()},
\code{\link{mm_inch}()},
\code{\link{mpa_kgf}()},
\code{\link{mpa_psi}()},
\code{\link{psi_mpa}()}
}
\concept{Measurement Unit Converter}
