% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fric_buzelli.R
\name{fric_buzzelli}
\alias{fric_buzzelli}
\title{Estimate pipe friction factor with Buzzelli formula}
\usage{
fric_buzzelli(reynolds, roughness = 0, strict = FALSE)
}
\arguments{
\item{reynolds}{\emph{Reynolds number}, []. Type: \code{\link[checkmate]{assert_double}}.}

\item{roughness}{relative roughness, []. Type: \code{\link[checkmate]{assert_double}}.}

\item{strict}{calculate only inside the precision region.
Type: \code{\link[checkmate]{assert_flag}}.}
}
\value{
Pipe friction factor, []. Type: \code{\link[checkmate]{assert_double}}.
}
\description{
Estimate \emph{Darcy friction factor} explicitly with extremely accurate
 \emph{Buzzelli} approximation of \emph{Colebrook equation}.
}
\details{
Buzzelli's formula is reported to be extremely accurate in the region:
 \itemize{
   \item \code{3.0e3 <= reynolds  <= 3.0e8}
   \item \code{    0 <= roughness <= 0.05}
 }

 In \code{strict = TRUE} mode argument values outside this precision region
 are not allowed, whereas in \code{strict = FALSE} either \emph{NA}s are
 generated in that case or calculation for laminar flow is performed when
 \code{reynolds < 2100.0}.
}
\examples{
library(pipenostics)

fric_buzzelli(c(2118517, 2000, 2118517), c(1e-6, 70e-3/1, 7e-3/1))

fric_buzzelli(c(2118517, 5500, 2118517), c(1e-6, 50e-3/1, 7e-3/1), TRUE)
}
\references{
\enumerate{
  \item  Offor, U. and Alabi, S. (2016) \emph{An Accurate and Computationally
         Efficient Explicit Friction Factor Model}. Advances in Chemical
         Engineering and Science, \emph{6}, pp. 237-245.
         \doi{10.4236/aces.2016.63024}.

  \item Buzzelli, D. (2008) \emph{Calculating friction in one step}. Machine
        Design, \emph{80} (12), pp. 54–55.
}
}
\seealso{
Other Fluid properties: 
\code{\link{fric_romeo}()},
\code{\link{fric_vatankhan}()},
\code{\link{re_u}()}
}
\concept{Fluid properties}
