% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fric_romeo.R
\name{fric_romeo}
\alias{fric_romeo}
\title{Estimate pipe friction factor with Romeo's formula}
\usage{
fric_romeo(reynolds, roughness = 0, strict = FALSE)
}
\arguments{
\item{reynolds}{Reynolds number, []. Type: \code{\link{assert_double}}.}

\item{roughness}{relative roughness, []. Type: \code{\link{assert_double}}.}

\item{strict}{calculate only inside the precision region. Type: \code{\link{assert_flag}}.}
}
\value{
pipe friction factor, []. Type: \code{\link{assert_double}}.
}
\description{
Estimate \emph{Darcy friction factor} explicitly with extremely accurate
 \emph{Romeo-Royo-Monzón} approximation of \emph{Colebrook equation}.
}
\details{
Romeo's formula is reported to be extremely accurate in the region:
 \itemize{
   \item \code{3.0e3 <= reynolds  <= 1.5e8}
   \item \code{0.0   <= roughness <= 0.05}
 }

 In \code{strict = TRUE} mode argument values outside this precision region
 are not allowed, whereas in \code{strict = FALSE} either \emph{NA}s are
 generated in that case or calculation for laminar flow is performed when
 \code{reynolds < 2100.0}.
}
\examples{
 library(pipenostics)

 fric_romeo(c(2118517, 2000, 2118517), c(0, 70e-3/1, 7e-3/1))
 # [1] 0.01028473 0.03200000 0.03373215  # []

 fric_romeo(c(2118517, 3030, 2118517), c(0, 50e-3/1, 7e-3/1), TRUE)
 # [1] 0.01028473 0.07859636 0.03373215  # []
}
\references{
\enumerate{
   \item Offor, U. and Alabi, S. (2016) \emph{An Accurate and Computationally Efficient Explicit Friction Factor Model}.
 Advances in Chemical Engineering and Science, \emph{6}, pp. 237-245. \doi{10.4236/aces.2016.63024}.

   \item Eva Romeo, Carlos Royo, Antonio Monzón, \emph{Improved explicit equations for estimation of friction factor in rough and smooth pipes},
 Chemical Engineering Journal, \emph{Volume 86}, Issue 3, 2002, Pages 369-374, ISSN 1385-8947. \doi{10.1016/S1385-8947(01)00254-6}.
 }
}
\seealso{
Other Fluid properties: 
\code{\link{fric_buzelli}()},
\code{\link{fric_vatankhan}()},
\code{\link{re_u}()}
}
\concept{Fluid properties}
