% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if97cptp1.R
\name{if97cptp1}
\alias{if97cptp1}
\title{IAPWS-IF97. Specific isobaric heat capacity of water}
\usage{
if97cptp1(temperature = c(300, 300, 500), pressure = c(3, 80, 3))
}
\arguments{
\item{temperature}{temperature of water, [\emph{K}]. Type: \code{\link{assert_double}}.}

\item{pressure}{pressure, [MPa]. Type: \code{\link{assert_double}}.}
}
\value{
Specific isobaric heat capacity of water, [kJ/kg/K], [kJ/kg/ºC].
 Type: \code{\link{assert_double}}.
}
\description{
Estimate specific isobaric heat capacity of water in accordance with
 \href{http://www.iapws.org/relguide/IF97-Rev.pdf}{IAPWS-IF97},
 \emph{Region 1}, as a function of temperature and pressure.
}
\details{
Even the provided function is vectorized its execution may be too slow for
 large datasets. In that case use specialized packages for estimating
 water-steam properties with IAPWS formulations.
}
\examples{
 # Typical usage in district heating:
 if97cptp1(temperature = k_c(65), pressure = mpa_kgf(6))
 # [1] 4.184094  # [kJ/kg/K] = [kJ/kg/ºC]
}
\references{
IAPWS Industrial Formulation 1997 for the Thermodynamic Properties of
  Water and Steam (\href{http://www.iapws.org/relguide/IF97-Rev.pdf}{IAPWS-IF97}).
}
\seealso{
Other Fluid properties: 
\code{\link{fric_buzelli}()},
\code{\link{fric_romeo}()},
\code{\link{fric_vatankhan}()},
\code{\link{if97pt4}()},
\code{\link{if97vtp1}()},
\code{\link{r12dv}()},
\code{\link{re_u}()}
}
\concept{Fluid properties}
