% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if97pt.R
\name{if97pt4}
\alias{if97pt4}
\title{IAPWS-IF97. Saturation pressure of water (steam)}
\usage{
if97pt4(temperature = c(300, 500, 600))
}
\arguments{
\item{temperature}{temperature of water (steam), [\emph{K}]. Type: \code{\link{assert_double}}.}
}
\value{
Saturation pressure of water (steam), [MPa].Type: \code{\link{assert_double}}.
}
\description{
Estimate saturation pressure of water (steam) in accordance with
 \href{http://www.iapws.org/relguide/IF97-Rev.pdf}{IAPWS-IF97},
 \emph{Region 4}, as a function of temperature.
}
\details{
Even the provided function is vectorized its execution may be too slow for
 large datasets. In that case use specialized packages for estimating
 water-steam properties with IAPWS formulations.
}
\examples{
 # Typical usage in district heating:
 if97pt4(temperature = k_c(65))
 # [1] 0.0250411  # [MPa]
}
\references{
IAPWS Industrial Formulation 1997 for the Thermodynamic Properties of
  Water and Steam (\href{http://www.iapws.org/relguide/IF97-Rev.pdf}{IAPWS-IF97}).
}
\seealso{
Other Fluid properties: 
\code{\link{fric_buzelli}()},
\code{\link{fric_romeo}()},
\code{\link{fric_vatankhan}()},
\code{\link{if97cptp1}()},
\code{\link{if97vtp1}()},
\code{\link{r12dv}()},
\code{\link{re_u}()}
}
\concept{Fluid properties}
