% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch.R
\name{inch}
\alias{inch}
\title{Millimeters to inches}
\usage{
inch(x)
}
\arguments{
\item{x}{numeric vector of lengths measured in \emph{millimeters} (\emph{mm})}
}
\value{
numeric vector of lengths measured in \emph{inches}
}
\description{
Convert length measured in \href{https://en.wikipedia.org/wiki/Millimetre}{millimeters} (mm)
 to \href{https://en.wikipedia.org/wiki/Inch}{inches}
}
\examples{
 inch(c(25.4, 1))
 # [1] 1.00000000 0.03937008  # [inch]

 ## unit test:
 stopifnot(round(inch(c(25.4, 1)), 8) == c(1.0, 0.03937008))

}
\seealso{
\code{\link{mm}} for converting \emph{inches} to \emph{mm}

Other utils: \code{\link{mm}}, \code{\link{mpa}},
  \code{\link{psi}}
}
\concept{utils}
