% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgf_mpa.R
\name{kgf_mpa}
\alias{kgf_mpa}
\title{Megapascals to kilogram-force per square}
\usage{
kgf_mpa(x)
}
\arguments{
\item{x}{pressure (stress) measured in \emph{megapascals}, [\emph{MPa}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
Pressure (stress) in
 \emph{kilogram-force per square cm}, [\emph{kgf/cm²}].
 Type: \code{\link[checkmate]{assert_double}}.
}
\description{
Convert pressure (stress) measured in
 \href{https://en.wikipedia.org/wiki/Pascal_(unit)}{megapascals} (MPa)
 to
 \href{https://en.wikipedia.org/wiki/Kilogram-force_per_square_centimetre}{
 kilogram-force per square cm} (\emph{kgf/cm²}).
}
\examples{
library(pipenostics)

kgf_mpa(c(0.0980665, 1))

}
\seealso{
\code{\link{mpa_kgf}} for converting \emph{kilogram-force per square cm} to
 \emph{megapascals}

Other Measurement Unit Converter: 
\code{\link{c_k}()},
\code{\link{f_k}()},
\code{\link{inch_mm}()},
\code{\link{k_c}()},
\code{\link{loss_flux}()},
\code{\link{mm_inch}()},
\code{\link{mpa_kgf}()},
\code{\link{mpa_psi}()},
\code{\link{psi_mpa}()}
}
\concept{Measurement Unit Converter}
