% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m278hlcha.R
\name{m278hlcha}
\alias{m278hlcha}
\title{Minenergo-278. Heat losses of pipeline segment in channel}
\usage{
m278hlcha(
  t1 = 110,
  t2 = 60,
  t0 = 5,
  insd1 = 0.1,
  insd2 = insd1,
  d1 = 0.25,
  d2 = d1,
  lambda1 = 0.09,
  lambda2 = 0.07,
  k1 = 1,
  k2 = k1,
  lambda0 = 1.74,
  z = 2,
  b = 0.5,
  h = 0.5,
  len = 1,
  duration = 1
)
}
\arguments{
\item{t1}{temperature of heat carrier (water) inside the supplying pipe, [\emph{°C}],
numeric vector.}

\item{t2}{temperature of heat carrier (water) inside the returning pipe, [\emph{°C}],
numeric vector.}

\item{t0}{temperature of environment, [\emph{°C}]. In case of channel laying this is
the temperature of subsoil, numeric vector.}

\item{insd1}{thickness of the insulator which covers the supplying pipe, [\emph{m}],
numeric vector.}

\item{insd2}{thickness of the insulator which covers the returning pipe, [\emph{m}],
numeric vector.}

\item{d1}{external diameter of supplying pipe, [\emph{m}], numeric vector.}

\item{d2}{external diameter of returning pipe, [\emph{m}], numeric vector.}

\item{lambda1}{thermal conductivity of insulator which covers the supplying pipe
[\emph{W/m/°C}], numeric vector.}

\item{lambda2}{thermal conductivity of insulator which covers the returning pipe
[\emph{W/m/°C}], numeric vector.}

\item{k1}{technical condition factor for insulator of supplying pipe, [], numeric vector.}

\item{k2}{technical condition factor for insulator of returning pipe, [], numeric vector.}

\item{lambda0}{thermal conductivity of environment, [\emph{W/m/°C}]. In case of channel
laying this is the thermal conductivity of subsoil, numeric vector.}

\item{z}{channel laying depth, [\emph{m}], numeric vector.}

\item{b}{channel width, [\emph{m}], numeric vector.}

\item{h}{channel height, [\emph{m}], numeric vector.}

\item{len}{length of pipeline segment, [\emph{m}], numeric vector.}

\item{duration}{duration of heat flux emittance, [\emph{hour}], numeric vector.}
}
\value{
Heat flux emitted by pipeline segment during \code{duration}, [\emph{kcal}].
 If \code{len} of pipeline segment is 1 \emph{m} and \code{duration} of
 heat flux emittance is set to 1 \emph{hour} then the return value is equal
 to that in [\emph{kcal/m/h}] units and so comparable with values of
 heat flux listed in
 \href{http://docs.cntd.ru/document/902148459}{Minenergo Order 325}.
}
\description{
Calculate values of heat flux emitted by pipeline segment mounted in channel
 as a function of construction, operation, and technical condition
 specifications according to
 Appendix 5.1 of \href{http://www.complexdoc.ru/ntdtext/547103/}{Minenergo Method 278}.

 This type of calculations is usually made on design stage of district
 heating network (where water is a heat carrier) and is closely related
 to building codes and regulations.
}
\details{
\code{k1} and \code{k2} factor values equal to one mean the best technical
  condition of insulation of appropriate pipes, whereas for poor technical
  state factor values tends to five or more.

  Nevertheless, when \code{k1} and \code{k2} both equal to one the calculated
  heat flux [\emph{kcal/m/h}] is sometimes higher than that listed in
  \href{http://docs.cntd.ru/document/902148459}{Minenergo Order 325}.
  One should consider that situation when choosing method for heat loss
  calculations.
}
\examples{
 ## Dummy test:
 stopifnot(round(m278hlcha(), 3) == 86.930)

 ## Naive way to find out technical state (factors k1 and k2) for pipe
 ## segments constructed in 1980:
 stopifnot(
   round(
     optim(
       par = c(1.5, 1.5),
       fn = function(x) {
         # functional to optimize
         abs(
           m278hlcha(k1 = x[1], k2 = x[2]) -
             m325nhl(year = 1980, laying = "channel", d = 250, temperature = 110)
         )
       },
       method = "L-BFGS-B",
       lower = 1.01, upper = 4.4
     )$par, 1) == c(4.3, 4.3) # c(k1, k2)
 )

}
\seealso{
Other Minenergo: 
\code{\link{m278hlair}()},
\code{\link{m278hlund}()},
\code{\link{m278insdata}},
\code{\link{m278inshcm}()},
\code{\link{m278soildata}},
\code{\link{m325beta}()},
\code{\link{m325nhldata}},
\code{\link{m325nhl}()},
\code{\link{m325tdrop}()}
}
\concept{Minenergo}
