% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m278insdata.R
\docType{data}
\name{m278insdata}
\alias{m278insdata}
\title{Minenergo-278. Heat conductivity factors of pipe insulation materials}
\format{
A data frame with 39 rows and 4 variables:
\describe{
  \item{id}{Identifier of insulation material according to table 5.1 of
    Appendix 5.3 in \href{http://www.complexdoc.ru/ntdtext/547103/}{Minenergo Method 278}.
  }
  \item{material}{Designation of insulation material more or less similar
  to those in table 5.1 of Appendix 5.3 in
  \href{http://www.complexdoc.ru/ntdtext/547103/}{Minenergo Method 278}.}
  \item{lambda}{Value for intercept, [mW/m/°C ]}
  \item{k}{Value for factor}
 }
}
\source{
\url{http://www.complexdoc.ru/ntdtext/547103/}
}
\usage{
m278insdata
}
\description{
Data represents values for calculating heat conductivity factor of pipe
insulation as a linear function of temperature of water as a heat carrier.
Those values are set for different insulation materials in
Appendix 5.3 of \href{http://www.complexdoc.ru/ntdtext/547103/}{Minenergo Method 278}
as norms.
}
\details{
Usually the data is not used directly. Instead use function \code{m278ins}.
}
\seealso{
Other Minenergo: 
\code{\link{m325beta}()},
\code{\link{m325nhldata}},
\code{\link{m325nhl}()}
}
\concept{Minenergo}
\keyword{datasets}
