% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Inches to mm}
\usage{
mm(x)
}
\arguments{
\item{x}{numeric vector of lengths measured in \emph{inches}}
}
\value{
numeric vector of lengths measured in \emph{millimeters} (\emph{mm})
}
\description{
Convert length measured in \href{https://en.wikipedia.org/wiki/Inch}{inches}
 to \href{https://en.wikipedia.org/wiki/Millimetre}{millimeters} (mm)
}
\examples{
 mm(c(0.03937008, 1))
 # [1]  1.0 25.4  # [mm]

 ## unit test:
 stopifnot(round(mm(c(0.03937008, 1)), 1) == c(1.0, 25.4))

}
\seealso{
\code{\link{inch}} for converting \emph{mm} to \emph{inches}

Other utils: \code{\link{inch}}, \code{\link{mpa}},
  \code{\link{psi}}
}
\concept{utils}
