% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_inch.R
\name{mm_inch}
\alias{mm_inch}
\title{Inches to mm}
\usage{
mm_inch(x)
}
\arguments{
\item{x}{numeric vector of lengths measured in \emph{inches}}
}
\value{
numeric vector of lengths measured in \emph{millimeters} (\emph{mm})
}
\description{
Convert length measured in \href{https://en.wikipedia.org/wiki/Inch}{inches}
 to \href{https://en.wikipedia.org/wiki/Millimetre}{millimeters} (mm)
}
\examples{
 mm_inch(c(0.03937008, 1))
 # [1]  1.0 25.4  # [mm]

 ## unit test:
 stopifnot(round(mm_inch(c(0.03937008, 1)), 1) == c(1.0, 25.4))

}
\seealso{
\code{\link{inch_mm}} for converting \emph{mm} to \emph{inches}

Other utils: 
\code{\link{inch_mm}()},
\code{\link{mpa_psi}()},
\code{\link{psi_mpa}()}
}
\concept{utils}
