% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_psi.R
\name{mpa_psi}
\alias{mpa_psi}
\title{Convert pounds per square inch to megapascals}
\usage{
mpa_psi(x)
}
\arguments{
\item{x}{numeric vector of pressure (stress) measured in \emph{pounds per square inch} (\emph{PSI})}
}
\value{
numeric vector of pressure (stress) measured in \emph{megapascals} (\emph{MPa})
}
\description{
Convert pressure (stress) measured in \href{https://en.wikipedia.org/wiki/Pounds_per_square_inch}{pounds per square inch} (PSI)
 to \href{https://en.wikipedia.org/wiki/Pascal_(unit)}{megapascals} (MPa)
}
\examples{
 mpa_psi(c(145.03773800721814, 1))
 # [1] 1.000000000 0.006894757 # [MPa]

 ## unit test:
 stopifnot(round(mpa_psi(c(145.03773800721814, 1)), 8) == c(1, 6.89476e-3))

}
\seealso{
\code{\link{psi_mpa}} for converting \emph{megapascals} to \emph{pounds per square inch}

Other utils: 
\code{\link{inch_mm}()},
\code{\link{mm_inch}()},
\code{\link{psi_mpa}()}
}
\concept{utils}
