% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipenostics-package.R
\docType{package}
\name{pipenostics-package}
\alias{pipenostics}
\alias{pipenostics-package}
\title{pipenostics: Diagnostics, Reliability and Predictive Maintenance of Pipeline Systems}
\description{
Functions representing some useful empirical and data-driven models of heat loss, corrosion diagnostics, reliability and predictive maintenance of pipeline systems. The package is an option for technical engineering departments of heat generating and heat transfer companies that use or plan to use regulatory calculations in their activities. Methods are described in Timashev et al. (2016) \doi{10.1007/978-3-319-25307-7}, A.C.Reddy (2017) \doi{10.1016/j.matpr.2017.07.081}, Minenergo (2008) \url{https://docs.cntd.ru/document/902148459}, Minenergo (2005) \url{https://docs.cntd.ru/document/1200035568}, Xing LU. (2014) \doi{10.1080/23744731.2016.1258371}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://omega1x.github.io/pipenostics/}
  \item Report bugs at \url{https://github.com/omega1x/pipenostics/issues}
}

}
\author{
\strong{Maintainer}: Yuri Possokhov \email{omega1x@gmail.com} (\href{https://orcid.org/0000-0002-3570-4337}{ORCID})

}
