% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipenostics.R
\docType{package}
\name{pipenostics}
\alias{pipenostics}
\title{Diagnostics, reliability and predictive maintenance of pipeline systems}
\description{
Functions representing some useful empirical and data-driven models of heat
losses, corrosion diagnostics, reliability and predictive maintenance of
pipeline systems. The package is an option for digital transformation of
technical engineering departments of heat generating and heat transferring
companies.
}
\details{
For further details visit the
\href{https://omega1x.github.io/pipenostics}{package site}.
}
