% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pof.R
\name{pof}
\alias{pof}
\title{Probability of failure of the corroded pipe}
\usage{
pof(
  depth = seq(0, 10, length.out = 100),
  l = seq(40, 50, length.out = 100),
  d = rep(762, 100),
  wth = rep(10, 100),
  strength = rep(358.5274, 100),
  pressure = rep(0.588, 100),
  temperature = rep(150, 100),
  rar = function(n) stats::runif(n, 0.01, 0.3)/365,
  ral = function(n) stats::runif(n, 0.01, 0.3)/365,
  days = 0,
  k = 0.8,
  method = "b31g",
  n = 1e+06
)
}
\arguments{
\item{depth}{maximum depth of the corroded area measured during \emph{inline inspection},
[\emph{mm}]. Type: [\code{\link{assert_double}}].}

\item{l}{maximum longitudinal length of corroded area measured during
\emph{inline inspection}, [\emph{mm}]. Type: [\code{\link{assert_double}}].}

\item{d}{nominal outside diameter of the pipe, [\emph{mm}]. Type: [\code{\link{assert_double}}].}

\item{wth}{nominal wall thickness of the pipe, [\emph{mm}]. Type: [\code{\link{assert_double}}].}

\item{strength}{one of the next characteristics of steel strength, [\emph{MPa}]:
\itemize{
  \item specified minimum yield of stress (\emph{SMYS})
        for use with \code{\link{b31gpf}} and \code{\link{b31gmodpf}}.
  \item ultimate tensile strength (\emph{UTS}) or specified minimum tensile
        strength (\emph{SMTS}) for use with other failure pressure codes
        (\code{\link{dnvpf}}, \code{\link{pcorrcpf}}, \code{\link{shell92pf}}).
}
Type: [\code{\link{assert_choice}}].}

\item{pressure}{\href{https://en.wikipedia.org/wiki/Pressure_measurement#Absolute}{absolute pressure}
of substance (i.e. heat carrier) inside the pipe measured near defect
position, [\emph{MPa}]. In most cases this is a nominal operating pressure.
Type: [\code{\link{assert_double}}].}

\item{temperature}{temperature of substance (i.e. heat carrier) inside the pipe measured near
defect position, [\emph{°C}]. In case of district heating network this is
usually a calculated value according to actual or normative thermal-hydraulic
regime. Type: [\code{\link{assert_double}}].}

\item{rar}{random number generator for simulating of distribution of radial corrosion
rate in pipe wall, [\emph{mm/day}]. The only
argument \code{n} of the function should be the number of observations to
generate. Type: [\code{\link{assert_function}}].}

\item{ral}{random number generator for simulating of distribution of longitudinal corrosion
rate in pipe wall, [\emph{mm/day}]. The only
argument \code{n} of the function should be the number of observations to
generate. Type: [\code{\link{assert_function}}].}

\item{days}{number of days that have passed after or preceded the \emph{inline inspection}, [].
Negative values are for retrospective assumptions whereas positives are for
failure prognosis. Type: [\code{\link{assert_int}}].}

\item{k}{alarm threshold for leakage failure. It usually \code{0.6}, \code{0.7}, or
\code{0.8}, []. If set to \code{1} no alarm before failure occurs.
Type: [\code{\link{assert_number}}].}

\item{method}{method for calculating failure pressure:
\itemize{
  \item \emph{b31g} - using \code{\link{b31gpf}}.
  \item \emph{b31gmod} - using \code{\link{b31gmodpf}}.
  \item \emph{dnv} - using \code{\link{dnvpf}}.
  \item \emph{pcorrc} - using \code{\link{pcorrcpf}}.
  \item \emph{shell92} - using \code{\link{shell92pf}}.
}
Type: [\code{\link{assert_choice}}].}

\item{n}{number of observations to generate for
\href{https://en.wikipedia.org/wiki/Monte_Carlo_method#Monte_Carlo_and_random_numbers}{Monte-Carlo simulations},
Type: [\code{\link{assert_count}}].}
}
\value{
Probability of pipe failure for each corroded area measured during
  \emph{inline inspection}. Type: [\code{\link{assert_double}}].
}
\description{
Calculate \emph{probability of failure} of the corroded pipe taking into
 account its actual level of defectiveness and exploiting
 \href{https://en.wikipedia.org/wiki/Monte_Carlo_method#Monte_Carlo_and_random_numbers}{Monte-Carlo simulation}.

 In the method they consider two possible failures for a single pipeline
 cross section with the on-surface and longitudinally oriented defect of the
 \emph{metal-loss} type:

 \describe{
    \item{\emph{rupture}}{a decrease of the value of failure pressure down to the
    operating pressure.}
    \item{\emph{leak}}{increase of the corrosion depth (defect) up to the specified
    ultimate permissible fraction of pipe wall thickness.}
 }

 They also assume change in depth and length of corrosion defects in time be
 close to linear.
}
\details{
Unlike
}
\examples{
pof()

}
