% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r12dv.R
\name{r12dv}
\alias{r12dv}
\title{IAPWS R12-08. Dynamic viscosity of water (steam)}
\usage{
r12dv(temperature, density)
}
\arguments{
\item{temperature}{temperature of water (steam), [\emph{K}]. Type: \code{\link{assert_double}}.}

\item{density}{density of water, [kg/m^3]. Type: \code{\link{assert_double}}.}
}
\value{
Dynamic viscosity of water (steam), [Pa*s]*1e-6.Type: \code{\link{assert_double}}.
}
\description{
Estimate dynamic (shear) viscosity of pure water substance over an extensive
 range of fluid states in accordance with
 \emph{Release on the IAPWS Formulation 2008 for the Viscosity of Ordinary Water Substance}
 (\href{http://www.iapws.org/relguide/visc.pdf}{IAPWS R12-08})
 as a function of temperature and density.
}
\details{
Even the provided function is vectorized its execution may be too slow for
 large datasets. In that case use specialized packages for estimating
 water-steam properties with IAPWS formulations.
}
\examples{
 # Typical usage in district heating:
 t <- 65 + 273.15
 rho <- 1/if97vtp1(t, mpa_kgf(6))
 r12dv(t, rho)
 # 433.0342  ## [Pa*s]*1e-6
}
\references{
\href{http://www.iapws.org/relguide/visc.pdf}{IAPWS R12-08}.
  \emph{Release on the IAPWS Formulation 2008 for the Viscosity of Ordinary Water Substance}
}
\seealso{
Other Fluid properties: 
\code{\link{fric_buzelli}()},
\code{\link{fric_romeo}()},
\code{\link{fric_vatankhan}()},
\code{\link{if97cptp1}()},
\code{\link{if97pt4}()},
\code{\link{if97vtp1}()},
\code{\link{re_u}()}
}
\concept{Fluid properties}
