% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strderate.R
\name{strderate}
\alias{strderate}
\title{DNV-RP-F101. De-rate yield stress and tensile strength of pipe due to
 temperature}
\usage{
strderate(x, temperature = 24.3)
}
\arguments{
\item{x}{specified minimum yield of stress (\emph{SMYS}), or ultimate tensile
strength (\emph{UTS}), or specified minimum tensile strength (\emph{SMTS})
as a characteristic of steel strength \strong{at room temperature},
[\emph{MPa}]. Type: \code{\link[checkmate]{assert_double}}.}

\item{temperature}{temperature of pipe wall, [\emph{°C}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
De-rated value of \emph{x}, i.e. of appropriate pipe material property,
  [\emph{MPa}] .
  Type: \code{\link[checkmate]{assert_double}}.
}
\description{
Temperature is highly influence on pipe material properties and especially
  on its strength. Since in
  \href{https://law.resource.org/pub/us/cfr/ibr/002/api.5l.2004.pdf}{API
   SPECIFICATION 5L} values of \emph{SMYS} or \emph{UTS} are postulated at
  room conditions, in case of higher temperature magnitudes they should be
  corrected. For that purpose
  \href{https://www.dnv.com/oilgas/download/dnv-rp-f101-corroded-pipelines/}{
   DNV-RP-F101} offers linear de-rating for \emph{SMYS} or \emph{UTS}
  according to figure 2-3.
}
\examples{
library(pipenostics)

pipe_specs <- api5l3tdata[api5l3tdata$origin == 10,]

# De-rate SMYS for a single value of temperature:
with(pipe_specs, strderate(mpa_psi(smys), 53))

# De-rate UTS for a set of temperature values:
with(
  pipe_specs,
  strderate(mpa_psi(uts),seq(0, 250, length.out = length(uts)))
)

}
\concept{DNV-RP-F101}
