% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b36dwthv.R
\name{b36dwthv}
\alias{b36dwthv}
\title{Verify the alignment between nominal pipe diameter and wall thickness}
\usage{
b36dwthv(d, wth)
}
\arguments{
\item{d}{nominal (outside) diameter of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{wth}{nominal wall thickness of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
Verification status.Type: \code{\link[checkmate]{assert_logical}}.
}
\description{
Check the correspondence inside nominal diameter/wall thickness value pair
 for pipes manufactured in accordance with regulatory standards.

 Verification is performed using reference values from the
 \code{\link{b36pipedata}}-table.
}
\examples{
 library(pipenostics)

}
\seealso{
Other utils: 
\code{\link{b36mass}()},
\code{\link{b36pipedata}},
\code{\link{geoarea}()},
\code{\link{meteos}()},
\code{\link{mgtdhid}()}
}
\concept{utils}
