% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m325nhl.R
\name{m325nhl}
\alias{m325nhl}
\title{Minenergo-325. Calculate normative heat losses of pipe}
\usage{
m325nhl(
  year = 1986,
  laying = "underground",
  exp5k = TRUE,
  insulation = 0,
  d = 700,
  temperature = 110,
  len = 1,
  duration = 1,
  extra = 2
)
}
\arguments{
\item{year}{year when the pipe is put in operation after laying or total overhaul,
numeric vector.}

\item{laying}{type of pipe laying depicting the position of pipe in space:
\itemize{
  \item \code{air},
  \item \code{channel},
  \item \code{room},
  \item \code{tunnel},
  \item \code{underground},
}
character vector.}

\item{exp5k}{pipe regime flag: is pipe operated more that 5000 hours per year? Logical
vector.}

\item{insulation}{insulation that covers the exterior of pipe:
\describe{
  \item{\code{0}}{no insulation}
  \item{\code{1}}{foamed polyurethane or analogue}
  \item{\code{2}}{polymer concrete}
},
numeric vector.}

\item{d}{internal diameter of pipe, [\emph{mm}], numeric vector.}

\item{temperature}{temperature of heat carrier (water) inside the pipe, [\emph{°C}],
numeric vector.}

\item{len}{length of pipe, [\emph{m}], numeric vector.}

\item{duration}{duration of heat flux emittance, [\emph{hour}], numeric vector.}

\item{extra}{number of points used for temperature extrapolation, single value.}
}
\value{
Heat flux emitted by pipe during \code{duration}, [\emph{kcal}].
 If \code{len} of pipe is 1 \emph{m} and \code{duration} of heat flux
 emittance is set to 1 \emph{hour} then the return value is in the same
 units as value of heat flux, [\emph{kcal/m/h}], accepted by
 \href{http://docs.cntd.ru/document/902148459}{Minenergo Order 325}.
}
\description{
Calculate normative values of heat flux that is legally affirmed by
 \href{http://docs.cntd.ru/document/902148459}{Minenergo Order 325} to be
 emitted by steel pipe of district heating system with water as a heat carrier.
}
\details{
Temperature extrapolation and pipe diameter interpolation are leveraged
 for better accuracy. Both are linear as it dictated by
 \href{http://docs.cntd.ru/document/902148459}{Minenergo Order 325}.
 Nevertheless, one could control the extrapolation behavior by \code{extra}
 argument: use lower values of \code{extra} for soft curvature near extrapolation
 edges, and higher values for more physically reasoned behavior in far regions
 of extrapolation.
}
\examples{
with(m325nhldata, {

 ## Dummy test:
 foo <- m325nhl(epoch, laying, exp5k, insulation, as.double(diameter), temperature)
 stopifnot(all(flux == foo))
 cat("\nDummy test: OK.")

 ## Linear extrapolation adopted in Minenergo's Order 325 using last 2 points:
 temperature <- seq(0, 270, 10)  # [°C]
 flux <- m325nhl(1980, "underground", TRUE, 0, 70, temperature)  # [kcal/m/h]
 plot(temperature, flux, type = "b")
 cat("\nGraphic test: OK.")
})

}
\seealso{
Other Minenergo: 
\code{\link{m325nhldata}}
}
\concept{Minenergo}
