% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m325nvl.R
\name{m325nvl}
\alias{m325nvl}
\alias{m325nml}
\title{Minenergo-325. Normative material loss of pipe}
\usage{
m325nvl(a = 0, d = 720, wth = 12, len = 1)

m325nml(
  temperature = 130,
  pressure = mpa_kgf(6),
  a = 0,
  d = 720,
  wth = 12,
  len = 1
)
}
\arguments{
\item{a}{heat carrier (water) volume loss factor of cylindrical pipe,
[\emph{h⁻¹}]. Type: \code{\link[checkmate]{assert_double}}.}

\item{d}{nominal (outside) diameter of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{wth}{nominal wall thickness of pipe, [\emph{mm}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{len}{pipe length, [\emph{m}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{temperature}{temperature of heat carrier (water) inside the pipe, [\emph{°C}].
Type: \code{\link[checkmate]{assert_double}}.}

\item{pressure}{\href{https://en.wikipedia.org/wiki/Pressure_measurement#Absolute}{absolute
pressure} of heat carrier (water) inside the pipe, [\emph{MPa}].
Type: \code{\link[checkmate]{assert_double}}.}
}
\value{
\describe{
   \item{
     For \code{m325nvl}}{volume loss of heat carrier per hour, [\emph{m³/h}].
   }
   \item{
     For \code{m325nml}}{mass loss of heat carrier per hour, [\emph{ton/h}].
   }
 }
 Type: \code{\link[checkmate]{assert_double}}.
}
\description{
Calculate normative material loss of heat carrier (water) from
 cylindrical pipe according to rule \emph{10.1.2} of
 \href{https://docs.cntd.ru/document/902148459}{Minenergo Order 325}.
}
\details{
The calculations are based on the \emph{a}-factor, which may be treated as
 the maximum allowed rate of heat carrier (water) volume loss per hour. So,
 its value varies from \emph{0.0} \emph{h⁻¹} (no loss of heat carrier)
 up to \emph{0.0025} \emph{h⁻¹} (maximum possible loss allowed). The
 cylindrical form of pipe is always assumed in calculations.
}
\examples{
library(pipenostics)

# According to Minenergo-325 it may be granted to loose right up to the next
# value of tons of heat carrier (water) per year (nine-month heating season)
# for the typical supplying 100-meter length pipe:
m325nml(a = 0.0025, len = 100) * 24 * 90

}
\seealso{
Other Minenergo: 
\code{\link{m278hlair}()},
\code{\link{m278hlcha}()},
\code{\link{m278hlund}()},
\code{\link{m278insdata}},
\code{\link{m278inshcm}()},
\code{\link{m278soildata}},
\code{\link{m325beta}()},
\code{\link{m325nhl}()},
\code{\link{m325nhldata}},
\code{\link{m325nxdata}}
}
\concept{Minenergo}
